fix transfer of images from forum
For future reference: this is due to the change in format in Discourse.
Images used to appear in the Discourse markdown with a URL starting with
/upload/. Now it's upload://, and a bit more obscure as to what's going
on. Best to right-click on the image in the preview immediately after
dropping the image in the post, then 'Open in new tab', and copy/paste
the URL from the /upload/ part.

In due course, we may need better methods to pick this up in the script,
but it's going to be messy...